Implement AccessToken and refreshing
I was wrongly using OAuth2::AccessToken with OpenIDConnect, even though
the two are not compatible (OIDC builds on rack-oauth2, not oauth).

The OIDC gem does not support refreshing tokens out of the box, nor
tracking when a token expires. This necessitates adding our own
implementation of access tokens so that we can serialize this data in
the session and refresh when the token expires.